requirements.txt
mne
numpy
pandas
PyYAML
scikit-learn
scipy
matplotlib
eegdash